+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This file holds the Changelog of the project and all relevant changes.
 - Implemented a console printer for formatted output per telegram
 - Implemented TelegramPlayer for replaying telegrams from files
 - First base information for GSM status
-- Implemented SerialSimulator
+- Implemented SerialSimulator to replay data on the serial port
+- Introduced interface IUserVisualizeable and LogPrinter, so the output can be selected by the user
 
 ## 1.0.0
 Initial release including support for live parsing and parsing of stored files.